Moved FX navigateBreadthFirst method to new FXUtil - using @CompileStatic and improved navigation algorithm
